Set global address space plan for create address buttons
